Update microsoft-word-quiz.md (#218)
Q21. Options is definitely wrong. Unsure as it's actually File > Open > Recent > Recover Unsaved Documents in Word. But there's no such option
Q22. A 3D model is something you can insert from the Insert tab while a sketch is not.
Q24. I think this should be done in the AutoCorrect Options as you want /mt to be replaced with "Microsoft Teams" as you type, not a single once off replacement.