Update Linux Questions (4 new ones) (#3025)
* Added git question #Q99

Q99. How does the -p option change the behaviour of the git add command

* Update Linux question #33 (Correction)

There were multiple right answers (one was misspelled).

* Update linux-assesment.md

~ moved unanswere questions down (#Q73 -> #Q77, #Q74 -> #Q78, #Q75 -> #Q79, #Q76 -> #Q80)

+ added questions with correct answer #Q73, #Q74, #Q75, #Q76, and the references.
I'm not 100% sure about the correct answer of #Q76, but since by default journald stores locally (so not remotely - answer 1); it stores logs in binary format (so not plain text - answer 2); and its logs are by default persistent (so it can't be "not persistent" - answer 4), I would say it's the third one. Also I got top 5%, maybe that "adds more points" to the probability of it being the correct answer